requirements.txt
Pandas
Numpy
python-dotenv
mysql-connector-python
pymysql
scikit-learn
Seaborn
catboost
matplotlib
xgboost
Flask
mlflow
dill